pr comment: move utxo types inside e2etest/
